Move carousel to display inside KeyboardAvoidingView
